moved test stream to external page served by test container
